added encrypt function to create encrypted documents

+++ encryption.py
@@ -1,27 +1,46 @@
+import os.path
 import shutil
 from cryptography.fernet import Fernet
 
 class Encryption():
-    def __init__(self, filename, key, document_identifier):
+    def __init__(self, filename, document_identifier):
         self.filename = filename
-        self.key = key
         self.document_identifier = document_identifier
 
         # check if key exists
-        self.doesKeyExist = False
+        self.doesKeyExist = True
 
     def write_key(self):
         """
         Generates a key and save it into a file
         """
-        key = Fernet.generate_key()
-        with open("key.key", "wb") as key_file:
-            key_file.write(key)
+        if self.doesKeyExist == False:
+            key = Fernet.generate_key()
+            with open("key.key", "wb") as key_file:
+                key_file.write(key)
+            self.doesKeyExist = True
 
-    def encrypt(self, filename, key):
-        f = Fernet(self.key)
-        with open(filename, 'rb') as file:
-            encrypted_data = file.read()
+    def load_key(self):
+        """
+            Loads the key from the current directory named `key.key`
+            """
+        return open("key.key", "rb").read()
 
-        with open('document_identifier', 'wb') as file:
+    def encrypt(self):
+        key = self.load_key()
+        print(key)
+        f = Fernet(key)
+        with open(self.filename, 'rb') as file:
+            plaintext_data = file.read()
+            encrypted_data = f.encrypt(plaintext_data)
+            file.close()
+
+            path = 'C:\Projects\secure index\Encrypted Files'
+            file_name = self.filename
+            print(file_name)
+            complete_path_name = os.path.join(path, file_name)
+
+        # Creates encrypted documents
+        with open(self.document_identifier + '.txt', 'wb') as file:
             file.write(encrypted_data)
+            file.close()
